Fix last commit, reorganize, add SaffronCiel skin to selector
