+++ Assembly-CSharp.LadybugTestMod.mm/Assembly-CSharp.LadybugTestMod.mm.csproj
@@ -72,6 +72,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Base.cs" />
+	<Compile Include="PlayerSelect.cs" />
     <Compile Include="Utils.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
