+++ Assembly-CSharp.LadybugTestMod.mm/Base.cs
@@ -149,76 +149,5 @@ namespace Assembly_CSharp
         }
     }
 
-    //machinespray messing around with a character selector
-    [MonoModPatch("global::OptionCtrl")]
-    public class patch_RevaDisplay : OptionCtrl
-    {
-        //Store selection information through reloads
-        private static int option = 0;
-        public static string heroID = "SaffronDemo";
-        public static string heroAnim = "Saffron";
-        public static string heroName = "Saffron";
-        [MonoModPublic]
-        private ItemManager itemMan;
-        [MonoModPublic]
-        private BC ctrl;
-
-        [MonoModReplace]
-        public void ClickEnableReva()
-        {
-            option++;
-            if (option > Utils.getCharacters().Length - 1)
-            {
-                option = 0;
-            }
-            this.UpdateRevaDisplay();
-        }
-        [MonoModReplace]
-        private void UpdateOutfitDisplay()
-        {
-        }
-
-        [MonoModReplace]
-        public void UpdateRevaDisplay()
-        {
-            //NPE Prevention
-            bool flag = ctrl.runCtrl.spCtrl == null;
-            revaDisplay.updateMode = AnimatorUpdateMode.UnscaledTime;
-            if (flag)
-            {
-                revaText.text = heroName;
-                revaDisplay.runtimeAnimatorController = this.itemMan.GetAnim(heroAnim);
-                return;
-            }
-            BeingObject hero = ctrl.runCtrl.spCtrl.heroDictionary[Utils.getCharacters()[option]];
-            revaDisplay.runtimeAnimatorController = this.itemMan.GetAnim(hero.animName);
-            revaText.text = hero.nameString;
-            heroID = hero.beingID;
-            heroAnim = hero.animName;
-            heroName = hero.nameString;
-        }
-    }
-    [MonoModPatch("global::SpawnCtrl")]
-    public class patch_SpawnCtrl : SpawnCtrl
-    {
-        public void CreatePlayers()
-        {
-            if (S.I.AUTOMATION)
-            {
-                Player component = this.PlaceBeing(patch_RevaDisplay.heroID, TI.tileGrid[this.playerSpawnPos[0].x, this.playerSpawnPos[0].y], 0, false).GetComponent<Player>();
-                component.gameObject.GetComponent<Animator>().runtimeAnimatorController = this.GetAnim(patch_RevaDisplay.heroAnim);
-                component.StartCoroutine(component.TestLoop());
-            }
-            else
-            {
-                Being hero = PlaceBeing(patch_RevaDisplay.heroID, TI.tileGrid[this.playerSpawnPos[0].x, this.playerSpawnPos[0].y], 0, false);
-                hero.gameObject.GetComponent<Animator>().runtimeAnimatorController = this.GetAnim(patch_RevaDisplay.heroAnim);
-            }
-            if (S.I.FAIRY)
-            {
-                this.CreateFairy();
-            }
-        }
-
-    }
+    
 }
